+++ packages/websocket-plugin/index.ts
@@ -1,0 +1,2 @@
+export { NgxsReduxDevtoolsPluginModule } from './src/devtools.module';
+export { NgxsReduxDevtoolsPlugin } from './src/devtools.plugin';